Moves axiosConfig to a specific folder. Small changes on teams/create. Remove file
