+++ frontend/src/components/communityMon/create.jsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import { useUser } from "../../userContext";
 import { useTranslation } from "react-i18next";
-import axiosInstanceCommunitymon from "../../axiosConfigCommunitymon";
+import axiosInstanceCommunitymon from "../../axiosConfigs/axiosConfigCommunitymon";
 import { useNavigate } from "react-router-dom";
 
 function CreateCommunityMon() {
