+++ frontend/src/components/pokeplay/guess.jsx
@@ -3,7 +3,7 @@ import { useTranslation } from "react-i18next";
 import Loading from "../Loading";
 import confetti from "canvas-confetti";
 import axios from "axios";
-import axiosInstance from '../../axiosConfig';
+import axiosInstance from '../../axiosConfigs/axiosConfig';
 
 import FullPokemonSelect from "../select";
