+++ frontend/src/components/poketools/EV.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import { useTranslation } from "react-i18next";
 
 import Loading from "../Loading";
-import axiosInstance from '../../axiosConfig';
+import axiosInstance from '../../axiosConfigs/axiosConfig';
 import Select from "react-select";
 import Error from "../error";
