+++ frontend/src/components/select.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import Select from "react-select";
 import { useTranslation } from "react-i18next";
-import axiosInstance from "../axiosConfig";
+import axiosInstance from '../../axiosConfigs/axiosConfig';
 
 function FullPokemonSelect({onChange}) {
   const [allPokemon, setAllPokemon] = useState([]);
